% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bins.R
\name{make_bins}
\alias{make_bins}
\title{takes a list of KNA data.tables and jointly bins them}
\usage{
make_bins(X, l = 50, C = c("PC1", "PC2", "NNmean"))
}
\arguments{
\item{X}{A list of KNA data.tables, the output of make_KNA}

\item{l}{The number of intervals to divide each dimension into}

\item{C}{A vector of column names that will be binned. Recommended to be first two SNPCA components and mean nearest neighbour distance}
}
\value{
Each KNA table is augmented with a bin ID column, the bin IDs are common across the images.
}
\description{
takes a list of KNA data.tables and jointly bins them
}
