% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_freq_cluster.R
\name{make_freq_cluster}
\alias{make_freq_cluster}
\title{takes a list of binned KNA data.tables and the frequency table, and makes frequency cluster}
\usage{
make_freq_cluster(X, tab, nclus = 4, split_clusters = TRUE, NNmean_break = 3)
}
\arguments{
\item{X}{list of binned KNA data.tables}

\item{tab}{the frequency table}

\item{nclus}{The number of clusters to be found in the frequency table. Currently hard coded to use hierarchical clustering with average linkage.}

\item{split_clusters}{There is often a 'leftover' cluster of bins that aren't populated over all the images, and these occur over a wide range of mean nearest neighbour values. This logical value indicates whether to split this cluster}

\item{NNmean_break}{The mean nearest neighbour distance to split the leftover cluster on}
}
\value{
The list of binned KNA data.tables is augmented with a new column indicating cluster membership
}
\description{
takes a list of binned KNA data.tables and the frequency table, and makes frequency cluster
}
