% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_frequency_table.R
\name{make_frequency_table}
\alias{make_frequency_table}
\title{takes a list of binned KNA data.tables and makes a frequency table}
\usage{
make_frequency_table(X)
}
\arguments{
\item{X}{a list of binned KNA data.tables}
}
\value{
a frequency table, where each row corresponds to an image. Each row sums to one.
}
\description{
takes a list of binned KNA data.tables and makes a frequency table
}
