% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_SNPCA.R
\name{make_SNPCA}
\alias{make_SNPCA}
\title{Calculates SNPCA from spatially random data}
\usage{
make_SNPCA(K, npoints = 1e+05, winsize = 2, border = 0.5)
}
\arguments{
\item{K}{The number of nearest neighbours, e.g. K = 100}

\item{npoints}{The number of spatially random points that are simulated}

\item{winsize}{The size of the window in which to simulate spatially random points}

\item{border}{The width of the border, points in which do not contribute to the SNPCA}
}
\value{
An object of class prcomp (from R Stats package) containing a spatially random SNPCA, to be applied to a set of SMLM images
}
\description{
calculates spatial neighbourhood PCA (SNPCA) that is based on spatially random data, to be applied to the images
}
